AD7680B1

+++ src/communication.h
@@ -4,7 +4,7 @@
 #include "config.h"
 
 void initCommunication();
-void addDataToBuffer(uint16_t adc_data, uint32_t afe_data);
+void addDataToBuffer(uint16_t adc_data, uint16_t afe_data);
 void sendBufferIfFull();
 void addDataToBufferSingle(uint32_t adc_data);
 void sendBufferIfFullSingle();
